requirements.txt
transformers==4.46.3
tokenizers==0.20.3
PyMuPDF
img2pdf
einops
easydict
addict 
Pillow
numpy
psycopg2-binary
pyyaml
sqlalchemy[asyncio]
asyncpg